# docs/yggdrasil/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ratatoskr WebPanel(Yggdrasil) Rest API",
        "title": "Ratatoskr(Yggdrasil)",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "ali.kaviani@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5050",
    "basePath": "/",
    "paths": {
...
        "/v1/auth/login": {
            "post": {
                "description": "Login user using email and password combination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user using email and password combination",
                "operationId": "handle_login",
                "parameters": [
                    {
                        "description": "Login Request Payload",
                        "name": "EventMessage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success Result",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "401": {
                        "description": "invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "403": {
                        "description": "account is not active or email is not confirmed",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            }
        },
        "/v1/auth/oauth/{provider}": {
            "get": {
                "description": "This endpoint generate an URL for the specified provider name on the url",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "operationId": "handle_oauth_step1",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth provider name (google,..)",
                        "name": "provider",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.OauthRedirectResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Validation error(s)",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.StandardResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/auth/signup": {
            "post": {
                "description": "Creates new account using email, password and company name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Create new account using basic account information",
                "operationId": "handle_signup",
                "parameters": [
                    {
                        "description": "Signup Request Payload",
                        "name": "EventMessage",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.SignupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success Result",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "409": {
                        "description": "Email is already in-use",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.StandardResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "ali.kaviani@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "testpassword"
                }
            }
        },
        "handlers.OauthRedirectResponse": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        },
        "handlers.SignupRequest": {
            "type": "object",
            "required": [
                "company",
                "email",
                "password"
            ],
            "properties": {
                "company": {
                    "type": "string",
                    "example": "ACME"
                },
                "email": {
                    "type": "string",
                    "example": "ali.kaviani@gmail.com"
                },
                "password": {
                    "type": "string",
                    "example": "testpassword"
                }
            }
        },
        "rest.StandardResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer",
                    "example": 0
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerToken": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}